認可

+++ pages/api/item/delete/[id].js
@@ -5,6 +5,9 @@ import auth from '../../user/auth.js'
 const deleteItem = async(req, res) => {
   try {
     await connectDB()
+    const singleItem = await ItemModel.findById(req.query.id)
+    if (singleItem.email !== req.body.email) throw new Error();
+    
     await ItemModel.findByIdAndDelete(req.query.id)
     return res.status(200).send({message: "delete complete"})
